Fixed errors create by merging.
