+++ PlanAndRide/PlanAndRide.BusinessLogic/Domain/Ride.cs
@@ -1,16 +1,20 @@
-﻿namespace PlanAndRide.BusinessLogic
+﻿
+using System.ComponentModel.DataAnnotations;
+
+namespace PlanAndRide.BusinessLogic
 {
     public class Ride
     {
-        public int Id { get; set; }
         [Required]
+        public int Id { get; set; }
+        
         public string Name { get; set; }
 
         public DateTime? Date { get; set; }
 
         public List<User> RideMembers;
 
-        public string Route { get; set; }
+        public  Route Route{ get; set; }
 
         public string Description { get; set; }
 
@@ -18,6 +22,6 @@
 
         public bool IsPrivate { get; set; }
 
-        public List<RouteRepository> Routes;
+        
     }
 }
