--- PlanAndRide/PlanAndRide.BusinessLogic/IRepository.cs.orig
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace PlanAndRide.BusinessLogic
-{
-    public interface IRepository<T> where T : class
-    {
-        T Get(int id);
-        IEnumerable<T> GetAll();
-        void Add(T entity);
-<<<<<<< HEAD
-        void Update(int id, T entity);
-        void Delete(int id);
-    }
-}
-=======
-        void Update(int id,T entity);
-        void Delete(int id);
-    }
-}
->>>>>>> main
