--- PlanAndRide/PlanAndRide.Web/PlanAndRide.Web.csproj.orig
@@ -1,26 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-<<<<<<< HEAD
-=======
-    <PackageReference Include="jQuery.Validation.Globalize" Version="1.1.0" />
->>>>>>> main
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="6.0.6" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\PlanAndRide.BusinessLogic\PlanAndRide.BusinessLogic.csproj" />
-<<<<<<< HEAD
-    <ProjectReference Include="..\PlanAndRide.GUI\PlanAndRide.GUI.csproj" />
-=======
->>>>>>> main
-  </ItemGroup>
-
-</Project>
