+++ PlanAndRide/PlanAndRide.Web/Program.cs.orig
@@ -1,23 +1,28 @@
 using PlanAndRide.BusinessLogic;
-<<<<<<< HEAD
-=======
-using PlanAndRide.Web.Models;
->>>>>>> main
+using PlanAndRide.Database.Repository;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddControllersWithViews();
 <<<<<<< HEAD
+builder.Services.AddSingleton<RouteRepository>();
 builder.Services.AddSingleton<RideRepository>();
 =======
-builder.Services.AddSingleton<RouteRepository>();
+
+builder.Services.AddSingleton<IRepository<PlanAndRide.BusinessLogic.Route>, RouteRepository>();
+builder.Services.AddSingleton<IRepository<Ride>, RideRepository>();
+
+builder.Services.AddScoped<IRouteService,RouteService>();
+builder.Services.AddScoped<IRideService,RideService>();
+
+
 
 >>>>>>> main
 var app = builder.Build();
 
 
-// Configure the HTTP request pipeline.
+// Configure the HTTP request pipeline.>
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Home/Error");
@@ -36,9 +41,6 @@ app.MapControllerRoute(
     name: "default",
     pattern: "{controller=Home}/{action=Index}/{id?}");
 
-<<<<<<< HEAD
-=======
-app.UseRequestLocalization("pl-PL");
->>>>>>> main
+app.UseRequestLocalization("en-US");
 
 app.Run();
