+++ PlanAndRide/PlanAndRide.sln.orig
@@ -1,0 +1,50 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32319.34
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlanAndRide.BusinessLogic", "PlanAndRide.BusinessLogic\PlanAndRide.BusinessLogic.csproj", "{70AD648A-E49D-42E2-B4F9-A6A95351C1AF}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlanAndRide.GUI", "PlanAndRide.GUI\PlanAndRide.GUI.csproj", "{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlanAndRide.Web", "PlanAndRide.Web\PlanAndRide.Web.csproj", "{B19EB418-BC55-44AB-9E7B-F34354A2E914}"
+<<<<<<< HEAD
+	ProjectSection(ProjectDependencies) = postProject
+		{70AD648A-E49D-42E2-B4F9-A6A95351C1AF} = {70AD648A-E49D-42E2-B4F9-A6A95351C1AF}
+		{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508} = {D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}
+	EndProjectSection
+=======
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlanAndRide.Database", "PlanAndRide.Database\PlanAndRide.Database.csproj", "{A0992AD9-1A0C-4E53-993D-86D73EBD8A80}"
+>>>>>>> main
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{70AD648A-E49D-42E2-B4F9-A6A95351C1AF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70AD648A-E49D-42E2-B4F9-A6A95351C1AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70AD648A-E49D-42E2-B4F9-A6A95351C1AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70AD648A-E49D-42E2-B4F9-A6A95351C1AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5C8A3D8-3D15-43F6-9AAC-8E4093DAB508}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B19EB418-BC55-44AB-9E7B-F34354A2E914}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B19EB418-BC55-44AB-9E7B-F34354A2E914}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B19EB418-BC55-44AB-9E7B-F34354A2E914}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B19EB418-BC55-44AB-9E7B-F34354A2E914}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A0992AD9-1A0C-4E53-993D-86D73EBD8A80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A0992AD9-1A0C-4E53-993D-86D73EBD8A80}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A0992AD9-1A0C-4E53-993D-86D73EBD8A80}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A0992AD9-1A0C-4E53-993D-86D73EBD8A80}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D9283903-275B-43A4-A1EC-62BFBA6C247B}
+	EndGlobalSection
+EndGlobal
